server/app/requirements.txt
# =========================
# server/app/requirements.txt
# =========================
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
starlette==0.38.4
httpx==0.27.0
websockets==12.0
numpy==2.0.1
soundfile==0.12.1
scipy==1.14.0
# STT
faster-whisper==1.0.3
# LLM local
llama-cpp-python==0.2.90
# OpenAI fallback (optional)
openai==1.43.0
# Embeddings (optional future RAG)
sentence-transformers==3.0.1
python-multipart==0.0.12
